JavaScriptCore:
        Reviewed by Kevin McCullough.

        Fourth round of fixes for implicit 64-32 bit conversion errors.
        <rdar://problem/5292262>

        Add custom piDouble and piFloat constants to use instead of M_PI.

        * kjs/math_object.cpp:
        (MathObjectImp::getValueProperty):
        * wtf/MathExtras.h:
        (wtf_atan2):

WebCore:

        Reviewed by Darin.

        Fifth round of fixes for implicit 64-32 bit conversion errors.
        <rdar://problem/5292262>

        Add toFloat() method to string classes and use it where appropriate.

        * WebCore.xcodeproj/project.pbxproj:
        * css/cssparser.cpp:
        (WebCore::CSSParser::lex):
        * ksvg2/svg/SVGAngle.cpp:
        (WebCore::SVGAngle::setValueAsString):
        * platform/AtomicString.h:
        (WebCore::AtomicString::toFloat):
        * platform/DeprecatedString.cpp:
        (WebCore::DeprecatedString::toFloat):
        * platform/DeprecatedString.h:
        * platform/FloatConversionUtilities.h: Added.
        (WebCore::narrowPrecisionToFloat):
        (WebCore::narrowPrecisionToCGFloat):
        * platform/PlatformString.h:
        * platform/String.cpp:
        (WebCore::String::toFloat):
        * platform/StringImpl.cpp:
        (WebCore::StringImpl::toFloat):
        * platform/StringImpl.h:



Canonical link: https://commits.webkit.org/18688@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@23931 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/css/cssparser.cpp
@@ -3012,7 +3012,7 @@ int CSSParser::lex(void* yylvalWithoutType)
         length--;
     case FLOAT:
     case INTEGER:
-        yylval->val = DeprecatedString((DeprecatedChar *)t, length).toDouble();
+        yylval->val = DeprecatedString((DeprecatedChar*)t, length).toFloat();
         break;
 
     default:
